bonus updated to mandatory state, cone is working except the base and norm vector

+++ miniRT/srcs_bonus/color_op_bonus.c
@@ -12,7 +12,7 @@
 
 #include "../includes/miniRT_bonus.h"
 
-t_pos	multconst_rgb(float c, t_pos color)
+t_pos	multconst_rgb(double c, t_pos color)
 {
 	t_pos	result;
